docs/aibot-sdk/sdks/python/source/aibot-python-sdk/requirements.txt
websockets>=12.0
aiohttp>=3.9
pyee>=11.0
cryptography>=42.0
certifi>=2023.0
python-dotenv>=1.0
